Table: improved/new implementation of WCS conversion
The old/initial implementation was very crude and hard to generalize. With
this commit, it can now work on any number of dimentions and also the
syntax has become the reverse polish notation, similar to Arithmetic. This
will allow easy inclusion of many operators in the near future.

Also, while doing this, I noticed that in the functions were we check file
formats based on file names/suffixes, we hadn't accounted for a
standard-input `filename' (which is NULL). This is now accounted for.