Arithmetic: no checking of size at read-time
Until now, when Arithmetic read a new dataset, it would check its size with
the reference dataset's size. However, because of operators like
`collapse-*', the size of the dataset may change. Therefore, if we read a
dataset that we want to change it size later, Arithmetic will complain and
abort.

All operators that work on more than one dataset check the sizes of the
inputs. Therefore, with this commit, we leave it to the operators to check
their inputs size and we won't check the size at read-time of the dataset.