Table: Column arithmetic now possible
With this commit, Table now has a new feature to do arithmetic operations
on various input columns. It is activated by starting the requested column
with `arith ' and can do almost all the operations in the Arithmetic
program, along with some other operators for coordinate conversion.

To use the same infra-structure as the Arithmetic program, some
generalizations were necessary in the Arithmetic library and programs also,
but now, along with Table, there is a much more clear and robust
implementation.