MakeCatalog: automatically find total number of clumps if necessary
Until now, MakeCatalog would use the `NUMLABS' keyword in the clumps
catalog to find the total number of clumps before parsing the dataset. This
was very inconvenient when we used clump labels that were not (directly)
created by Segment.

With this commit, when `NUMLABS' is not present, MakeCatalog will parse the
clumps and objects datasets and count the total number of clumps in all the
objects while also re-labeling the clumps dataset (and saving the relabeled
image).