Fits: --primaryimghdu allows copying/cutting to zero-th extension
Until now, it wasn't easy to copy a requested extension to the zero-th
extension of a FITS file. With this commit, the Fits program has a new
`----primaryimghdu' option which will copy the requested image HDU into the
zero-th extension of an output file that doesn't yet exist.

This task was suggested by Zahra Bagheri.