Programs: extra dimensions (length 1) are removed before processing
Until now, when the input dataset has length-1 dimensions, the programs
would crash with a message that the number of dimensions isn't
correct. With this commit, immediately after reading the dataset, all such
redundant dimensions are removed. Two library functions were added to allow
easily doing this.

This task was done on a dataset provided by Zahra Bagheri.