MakeCatalog: relabeled clumps image written in proper directory
Until now, even when an output file name was given, the relabeled clumps
image would be kept in the top level directory. This didn't correspond to
the current way Gnuastro programs work.

With this commit, when an output file is given, the relabeled clumps image
is stored in the same directory.