Fits: removed extra left-over comment
In a previous commit (784b220, "Fits: --primaryimghdu allows
copying/cutting to zero-th extension"), I had left-over a commented message
in `fits_hdu_copy' which I had put to help during its development! This
should have been removed before the commit, but I had forgot! It has been
removed with this commit.

Also, in the same function, I noticed that since `out' is now set within
the loop, it is cleaner to check if its not NULL before closing the file.