statistics library: properly accounting for maximum in histogram
Until now, the last element's count would be written into a patch of memory
right outside the allocated space for the histogram values. This has been
corrected with a better implementation of the counting step to account for
situations when floating point errors cause this situation.

This bug was reported by Raul Infante-Sainz.

This fixes bug #56480.