Installation: all libraries linked through -l
Recently we started using Gnulib's `AC_LIB_HAVE_LINKFLAGS' to search for
libraries and prepare the compilation flags. However, we have been having
some problems like [1], [2] or [3]. The core of this problem is that this
Gnulib macro actually returns the absolute address of the shared library.

With this commit, an Autoconf macro is added to fix this problem: when the
returned library is an absolute address, it will extract its directory and
add it as a `-L' to the LIBS and try linking with it through `-l'.

[1] https://lists.gnu.org/archive/html/bug-gnulib/2019-05/msg00088.html
[2] https://lists.gnu.org/archive/html/bug-gnulib/2019-05/msg00155.html
[3] https://debbugs.gnu.org/cgi/bugreport.cgi?bug=36184