Library (pointer.h): message when creating/deleting mmap'd file
Until now, the creation of memory-mapped files and their deletion was done
quietly in the background. This had two negative side-effects: 1) The
program would run much slower. 2) The left-over temporary files when the
program crashes would consume too much space (and the user wouldn't know to
delete them manually).

With this commit, the memory allocation function in charge of this
(`gal_pointer_allocate_mmap') accepts a new argument called
`quietmmap'. When it is zero, a message will be printed informing the user
of the file name and its size, and that they can use `--minmapsize' to
toggle this feature. Another message will later be printed when its
freed. This option then propagated to all the higher-level library
functions that call it and subsequently to the programs. As a result, a new
common option called `--quietmmap' is added to Gnuastro to let program
users directly disable such messages if it is annoying for them.

This task was implemented after a discussion with Michael Wilkinson.