Installation: updating LD_LIBRARY_PATH and setting rpath
On some systems, without setting rpath, the configure script won't find the
Gnulib components and will crash. On other systems, we need the libraries
to be linked in the proper order (so we don't want to give their absolute
filename, ending in `.so' for example). These were both problems we were
having because of the new `AC_LIB_HAVE_LINKFLAGS' feature.

This commit is an attempt to solve both problems: setting rpath and also
updating LD_LIBRARY_PATH for the tests.