Configuration: math library now a dependency of CFITSIO
On a Fedora 28, we noticed that CFITSIO fails because of not finding basic
math functions (even though `-lm' is in LDADD before CFITSIO). The solution
was to manually add `m' as a dependency during configuration.

Also, the most recent version of CFITSIO installs shared libraries by
default, so there is no need for the `make shared' rule.

The CFITSIO linking problem was found thanks to Hamed Altafi.