MakeProfiles: rounding of sub-pixel center at 1/1000000 from 1/100
Until now, to fix floating point issues (especially when converting from
WCS), we were rounding subpixel, image coordinates to 1/100. As a result,
for example a central pixel value of 2.499 would be rounded to
2.5. However, especially when using large oversampling, such levels of
precision are necessary. We are also dealing with double-precision floating
point types which are accurate to roughly 15 decimals. SO the rounding to
1/100 was not too accurate and could cause unexpected behavior.

To fix this problem, in MakeProfile's `oneprofile_center_oversampled'
function, the rounding is now set as a variable `r', and given a value of
1000000. This is also arbitrary and hard-coded in the source, but
currently, I don't see any need to change it. If we later feel the user
must be able to change it, we can define it as an option.

While inspecting the code, a few typos and edits were also fixed the
comments to make them more clear to read. Also, the name of the
`oneprof_set_prof_params' function was corrected to
`oneprofile_set_prof_params' (which is the correct format!).

This bug was reported by Roberto Baena Gallé.

This fixes bug #56641.