Warp: checking for option value existance before its contents
The `ui_add_to_modular_warps_ll' function was in charge of preparing the
individual warping parameters into a final matrix. Until now, it would
assume that all the functions have an argument (to make sure that its not
an empty string). But we had forgot to account for the `--align' option
that doesn't take an argument.

With this commit, the check for the contents of the argument is only done
when an argument is actually given.