Book: full edit of the general tutorial, broken into subsections now
Until now, the general program usage tutorial in the book was a little too
long (23 pages in PDF). This made it hard for new readers to follow the
relation between the major steps.

With this commit, it is broken up into subsections, and as part of doing
this, I also made some minor additions, edits and clarifications.