Low surface brightness tutorial fully edited and corrected
Since version 0.9, a major bug-fix happened in NoiseChisel and the old set
of commands in the tutorial weren't good any more. They also didn't
correspond to the text: causing confusion for the readers.

So I went through the tutorial and fully edited and corrected it to work
nicely with the new NoiseChisel. While doing this, I found out that the
single erosion we were doing to estimate the detection's outer surface
brightness was mainly just noise, just doing one more erosion drastically
brings it up from 1/20th of the noise level to 1/4th of the noise-level. So
we now say that we reach 1/4th of the noise level. The corresponding
magnitude/arcsec^2 was udpated to ~28.3.

Finally, since the Hubble tutorial can't be done using itself (we don't
introduce any datasets to work with), and because almost all of it is no
put in the general tutorial, it has been removed.

Also, the following to changes were made in the code of Gnuastro while
doing the edit of this tutorial:

  1- I noticed that the Arithmetic program's collapse operator is not
     checking the dimensions of the input with the dimension requested to
     collapse. So a check was added.

  2- NoiseChisel doesn't check for the number of neighbors before rejecting
     outliers. This didn't fit the logic and would only cause confusion:
     the number of reported tiles in the warning messages changed with
     non-relevant options.

This fixes bug #56635.