Installation: configure script checks before using -Wl,-R
Until now, for the internal processing of the configure script, we would
convert any `-R's in `LDADD' to `-Wl,-R'. But while testing on a macOS, we
noticed that it crashes and complains about not understanding this
option. The problem was fixed when we removed this conversion.

Therefore with this commit, we do a simple check before the conversion and
if the compiler doesn't recognize `-Wl,-R', we won't convert.

This fixes bug #56662.