qsort library: floating point NaN values placed in the end
Until now, all the `gal_qsort' functions that worked with floating point
types simply returned the standard check (like `(ta > tb) - (ta < tb)' for
an increasing sort). However, if any, or both, of the elements were NaN, it
would return 0 (meaning equal). This would create an incorrect sorted
result.

To fix the problem, with this commit a `COMPARE_FLOAT_POSTPROCESS' macro
has been added in `lib/qsort.c' and all floating point comparison functions
will use it when the result of the basic comparison is 0.

This bug was reported by Roberto Baena Gallé.

This fixes bug #56671.