Installation: configure script updates LIBS while testing libraries
Until now, we would only use the LTLIB<NAME> output of
`AC_LIB_HAVE_LINKFLAGS'. But that is for libtool, not basic builds during
the configure script, so we would have to correct that manually.

But `AC_LIB_HAVE_LINKFLAGS' also outputs LIB<NAME> which is directly for
the compiler (not through libtool). We are now using this instead.