Installation: new --without-libNAME options in configure script
Until now, if an optional dependency library was present on a computer,
Gnuastro would link with it anyway. But its an optional dependency, and
some users may not want to link with it when they don't use it. In
particular, the library may have run-time problems, causing all Gnuastro
programs to crash. This actually happened. It is thus beneficial build
Gnuastro without the problematic library, but there was no way to do this
without modifying the code.

With this commit, we now have the `--without-libNAME' options to disable
optional libraries at configure time and let Gnuastro build without them.

This was suggested by David Valls-Gabaud.