Segment: labeled pixels set to blank when input has blank values
Until now, when the input has blanks but the labeled image doesn't (or has
blanks in other places), the inconsistency between the two can cause a
crash, something like this:

astsegment: a bug! Please contact us at bug-gnuastro@gnu.org to fix it. 1
pixels of detection 1 have not been labeled (as an object)

This happens is because on the labeled image pixels may be connected, but
when accounting for the blank pixels of the input, they maybe separate
regions.

With this commit, after Segment reads the labeled image, if the input has
blank values, they are set to blank on the labeled image is well.

This bug was reported by Roberto Baena Gallé.

This fixes bug #56709.