NoiseChisel: binary image before growth has blank if input does
Until now, we weren't forcing the binary image before growth to have blank
values. Therefore in some scenarios it would keep blank values in the
output and in others it wouldn't.

With this commit, when the labeled image has blank values, the respective
pixel is also given a blank value in the binary image (which may later be
used as output).

This fixes bug #56710.