CosmicCalculator: check applied in values to --obsline option
Until now, if a single value was given to this option, we would get a
segmentation fault. This is because it wouldn't check if the option is
given more than one value.

With this commit, if no value is given after the first, CosmicCalculator
aborts with an informative error.

Since a small change was made in the library, I also incremented the
library soname version with this commit. Many more changes in the library
will be made before the next major release.

This fixes bug #56736.