All programs: cp->program_struct is now available in all programs
Until now, we were only defining this tructure for programs that actually
used it. But this causes confusion when trying to use `cp->program_struct'
in a program that hasn't yet defined it (for example the case of the next
commit!).

With this commit, it is now defined in all the programs.