ConvertType: corrected missing conversion in SLS colormap
Until now, we were missing the case where the checked fraction in the SLS
color map is `159', as a result, when an element's value fell in this
range, it would become black. This was a typo for the time I was writing
this step!

After checking with SAO DS9 (which has the original SLS color-map
definition), I noticed that the values are indeed in sequence (so I hadn't
missed any line). I had just miss-typed `159' with `160'.

This fixes bug #56747.