Library (statistics): properly accounting for equal values
Until now, when the sigma-clipped values were all equal, the
`gal_statistics_sigma_clip' function would return NaN values! This happened
because when all the elements are equal, their standard deviation is
zero. But we define the tolerance by division with the standard
deviation. As a result the tolerance would be infinity (and larger than the
user's tolerance level). Therefore it would go to the maximum internal
value and just return NaNs.

With this commit, we now explicity check for the case where the standard
deviation is zero and when it is, we use the current mean, median and STD
as the "old" values to return in the end.

This bug was reported by Raúl Infante Sainz.

This fixes bug #56754.