Book: inclusiveness of --range in Table is now described
Until now, I had forgot to write the inclusiveness of the values given to
the `--range' operator to the Table program. With this commit, it is
explicitly mentioned.

Also, some extra cosmetic spaces in the two calls to `gal_arithmetic' have
been removed, to be more clear. They were useful in the initial
implementation of this function when the third condition's elements would
also align, but that doesn't align any more, so these spaces were
redundant and just confusing.