Arithmetic: new --onedonstdout option to print 1D outputs on stdout
Until now, if you wanted to use the outputs of the `unique' operator for
example in your shell script, it was necessary to save it as a file, then
read that file which is annoying.

With this commit, Arithmetic has a new `--onedonstdout' operator for this:
when it is called, any 1D output will be written directly to the standard
output.