Table: descriptive error printed if sort column has string type
Until now, if the column given to Table's `--sort' option was a string
type, it would crash with a "bug-report" type message, giving a type number
that is not recognized. This is not understandable to the user. Note that
this can happen: some FITS tables actually have numbers that are printed as
strings!!!

With this commit, when the requested column is a string, a desctiptive
error message is printed, even suggesting ways to fix it.