doc/genauthors: now accounts for names with more than two components
Until now, the `doc/genauthors' script (which is in charge of updating the
list of authors in the PDF book) could only deal with two component
names. However, last commit's author (Miguel de Val-Borro) has a
three-component name, so it wasn't displayed properly.

With this commit, AWK actually loops through the tokens to print the
name, so it can now deal with names with any number of components.