doc/genauthors: won't link to .mailmap if it already exists
Until now, `doc/genauthors' was assuming that it is run within the
project's `doc' directory. However, it may happen (happened to me!) that it
is run within the top project directory. In this scenario, it will delete
the `.mailmap' file.

With this commit, it checks the file's existance (as a link or as a normal
file) and a link is only made (and later deleted) when it didn't already
exist.