Book: source in new convention of one sentence per line (half complete)
Until now, the book's source is written with the policy of lines that are
75 characters. Since the book is often read by many people and suggestions
to improve it are alot, it is often edited. But following the fixed-width
line convention would result in a whole paragraph changing because of
changing one word for example. This doesn't fit into the logic of Git too
nicely and thus doesn't make immediately clear what was changed in each
commit.

With this commit, the printed text of the book is now written in the
convension of sentence per line. In this way, when a change is made in a
sentence, only that sentence will be marked by Git and we'll have a much
better control of the book's history.

Almost half of the book has now been converted to this format, in due time,
we can complete the rest (before any next release).