Table: new --equal and --notequal operators to select specific rows
Until now, if you wanted to select some special rows by equality of their
value with something, you had to use the `--range' option and give a value
just larger than it (but not large enough to include another
row). Alternatively, there was no way to remove some rows by equality (for
example to discard some rows of the output based on their IDs). You would
have to write a relatively complext AWK statement for this job. Both are
very inconvenient and prone to errors.

With this commit, Table now has two new options `--equal' and
`--notequal'. They can be used for the two tasks mentioned above: selecting
rows based on equality with some given criteria. It is therefore very easy
now to select/discard some objects based on their IDs for example.

Since this is ultimately a "selection" of the rows (like `range'), the
implementation of `--range' was also modified to become generic and eaisly
be used for these two operators is well (only the final step changing), and
in the future, easily add new options for other operations.