Table: allocating names of columns to read for selection
In the recent commit where the selection operations of Table were
re-written, I had forgot to allocate space for the column identifiers when
it necessary to read them. We thus had a crash.

With this commit, the allocation flag is now set and the crash is fixed.
Also, I noticed that there was an extra redundant `;' in
`ui_check_select_sort_before' which is now removed.