CosmicCalculator: numbers followed by a space only when necessary
Until now, in any situation, a specific calculation of CosmicCalculator
would print a number followed by a space. This was mainly good/intended for
cases where multiple numbers are requested. However, when only a single
number is requested and that number is to be put into a variable to be used
in a script, the extra space in the end will cause problems.

With this commit, an extra space is only printed between the output
numbers, not after the last printed number. Therefore if only one number is
requested, the output string won't have any extra space characters.