BuildProgram: --cc option allows using custom C compiler
Until now `gcc' was hard-coded into BuildProgram, so a GCC that is in the
search path was the only way to run BuildProgram. This would cause a crash
cases when there is no `gcc' in the PATH.

With this commit, BuildProgram has the new `--cc' option to let the user
explicitly set the compiler. If `--cc' hasn't been used, then it will look
into the `CC' and `GCC' environment variables and ultimately, if it can't
find anything in them, it will use `gcc'. This gives the users much greater

This was reported by Sebastián Luna Valero, while attempting to package
Gnuastro for conda-forge.