Internal Library: directory write permission checked with temp file
Until now, we were using the `faccessat' function for checking the
permission flags to write in a library (in particular with the `AT_FDCWD'
and `AT_EACCESS' macros). But these two macros recently caused a crash on
macOS 10.9 (while trying to build for Conda).

To fix this problem and have a more generic test, we now actually try to
build a temporary file in the directory (and delete it afterwards). This
totally removes the need for the `faccessat' library function and thus the
need for those problematic macros on older machines.

This fixes bug #56999.