Library WCS: coordinate conversions check if WCS isn't NULL
Until now, there was no actual check in the two `gal_wcs_img_to_world' and
`gal_wcs_world_to_img' functions to see if the input WCS pointer actually
points to something! As a result, when the library user forgot to read the
input file's WCS, they would get a segmentation fault.

With this commit a check has been added and the calling program will abort
when the input WCS is NULL.