Statistics: --contour option calculates contours for PGFPlots
One of the common things we need when doing an 2D analysis of an image is a
contour plot. Plotting tools like Matplotlib do calculate it internally,
but others like the LaTeX PGFPlots package doesn't. PGFPlots doesn't need
all the dependencies of Matplotlib and is much more easier to build within
LaTeX and creates much higher quality plots.

With this option, it is now very easy to create the necessary inputs to
PGFPlots and use it to make contours over a dispalyed image.

To do this, a new function has also been added to the library:
`gal_binary_connected_indexs', it operates very similar to the connected
component labeling function, but just keeps the labels.