BuildProgram: reading LDFLAGS and CPPFLAGS env for customization
Until now, BuildProgram's compilation was only configured through options
(and just recently the `CC' and `GCC' environment variables). But in some
scenarios, the user may want better control of the build and the default
`-L' and `-I' options of BuildProgram (same as the linking and include
options of the C compiler) may not be enough.

We recently came up with such a scenario when trying to build Gnuastro
through Conda on macOS. It added several linking flags through `-Wl' that
BuildProgram ignores.

With this commit, by default BuildProgram will use the `CPPFLAGS' and
`LDFLAGS' environment variables also. But their checking can be disabled
with the new `--noenv' option.