Crop: image mode can now account for no overlap
Until now, when there was no overlap with the input image, in image-mode
Crop would crash. This was because no FITS pointer would be created.

With this commit, after defining the crop, if no FITS pointer is created,
it will simply switch the `numimg' variable to zero and Crop will operate
as expected.

This fixes bug #57101.