Book: added a section on writing scripts in the general tutorial
After various workshops that were held with the tutorial, I felt its
necessary to also introduce readers to writing scripts and best practices
in doing it. I will later add a script for the whole tutorial, and also add
a section in this tutorial on writing Makefiles as a more optimal way of
organizing the processing.