MakeCatalog: labels that aren't present in image won't be parsed
MakeCatalog works by defining tiles over the individual labels of the
dataset and only parsing those tiles for each label in parallel. Also,
MakeCatalog assumes all integers until the largest integer in the image are
a label. Therefore, when a given label doesn't exist in the image,
MakeCatalog should actually set a tile size of 0 elements.

But this wasn't happening: to find the tiles of each label, we were
effectively assuming the largest possible tile, then shrinking it to fit
over the given label. Therefore, when a label/integer didn't exist in an
image, its tile would be larger than the actual input image size and could
cause a segmentation fault.

With this commit, before defining the tiles, we make sure that their
regions have actually been set. If they haven't, the tile's size will be
set to zero.

This fixes bug #57164.