MakeCatalog: Rows with no label are not shown in output
Until now, MakeCatalog would keep a row for every integer from 1 to the
largest label in the input dataset. For outputs of programs like
NoiseChisel and Segment, this was fine (the labels are consecutive and
contiguous), but when the labeled image is created by some other scenario,
This was not good. In some situations, the labels could start from
arbitrary large numbers and as a result, there would be many blank rows in
a very large output table.

With this commit, all the rows/IDs that don't contain any pixels are
removed from the output catalog, making it much more directly useful.