MakeCatalog: better implementation of ignoring integers not in input
2 commits ago (0ebad2b4ef), this new feature was added to ignore creating a
final catalog row for integers that don't actually have a pixel in the
input. But that implementation was not very efficient: it involved defining
a 1-pixel sized tile to parse over the non-existant integers and also
allocating large space in memory for the extra output rows for them during
the processing (they were only removed, to not be printed in the output, in
the final step).

With this commit, a much better implementation is done where tiles (and
output rows) are only generated for labels that actually exist in the input
and no extra rows are created in the middle of the processing. This will
results in much faster processing and less memory consumption.