MakeCatalog: cpscorr is properly set when --instd is a number
Until now the `p->cpscorr' (to correct for counts-per-second datasets), was
only set when an input image was given to `--instd'. As a result, when a
single number was given, it would be zero and `1/p->cpscorr' would become
infinity and the `--sn' column would be `inf'. With this commit, we also
set `p->cpscorr' when the input is a single number.

This bug was reported by Joseph Putko.

This fixes bug #57180.