MakeCatalog: initialized variable to avoid compiler warning
With the recent commits in setting the index of of the object in the final
catalog, an "warning: ‘oind’ may be used uninitialized in this function
[-Wmaybe-uninitialized]" came up during the build.

With this commit, its initialized to the largest possible number. It will
be immediately corrected, but by any chance, if the object's ID isn't
found, the program crashes with a segmentation fault instead of giving
unreasonable values.