Installation: correct wcslib reference and optional libs in pkg-config
Until now, in gnuastro's pkg-config file, we were using `wcs' for WCSLIB
instead of the standard `wcslib' name that the library uses in its
installed pkg-config file (`wcslib.pc'). As suggested by Stefan Brüns, it
should be `wcslib'.

Also, while making the change, I noticed that we weren't mentioning the
optional libraries in Gnuastro's pkg-config at all! So a test was also
added to see which optional libraries exist and adds the ones that
Gnuastro's library was built with to the installed `gnuastro.pc'.

This fixes bug #57200.