Book: removed changes since publication in NoiseChisel and Segment
Until recently (prior to the publication of arXiv:1909.11230), these
sections were the only place to introduce the changes since publication of
NoiseChisel and Segment. But since that paper's publication, readers can
just go read that and having this separate section (for future changes)
could be confusing.

So we just recommend to read both the 2015 and 2019 papers, and later (when
there are changes that must be noted), we'll fill that section the things
that aren't in any of the papers.