BuildProgram: CPPFLAGS and LDFLAGS passed to libtool after own options
Since the first implementation of BuildProgram reading CPPFLAGS and LDFLAGS
(in commit `1d4043f333ef'), they were being passed onto libtool with higher
precedence than BuildPrograms own options. This could cause problems during
`make check' when an older version of Gnuastro exists, and the old
installation's `--includepath' is in CPPFLAGS (as it should be!).

With this commit, the values given to CPPFLAGS and LDFLAGS are passed to
libtool after the values given to BuildProgram's `--includedir' and
`--linkdir'. In this way, if they are indeed unique, they will be used, but
they won't cause unnecessary conflicts like the scenario above.

This bug was reported by Takashi Ichikawa.