Book: String column Sufi tutorial's catalog now set to 6 characters
Until now, the catalog of the Sufi tutorial had `str7' for the number of
characters in the 4th column. But this is based on the assumption that
there is atleaset two space characters after the `t' of `moffat' and before
the `5' of `5.000'. But when importing the catalog into a plain text
editor, readers may mistakenly only put one space there and thus the `5'
character will be missed and MakeProfiles will complain with a `0' value
for the radius column.

With this commit, the total number of characters in the string column has
been set to 6 which is also enough and will avoid such problems.

This was reported by Marcel Popescu.