MakeCatalog: correct treating of NaN values pixels
Until now, MakeCatalog's `p->hasblank' was mistakenly only set based on the
labels image, not the values image. Also, if there were NaN pixels in the
values image (and not the labels image), then the brightness and magnitude
error columns and the S/N column would become NaN.

With this commit, we now check the Sky value and the Sky standard deviation
value to be NaN before actually using it. It was thus necessary to keep a
count of how many pixels go into each. Also, when estimating `CCOL_SUM_VAR'
and `OCOL_SUM_VAR' (the sum of the variances including the values dataset)
only non-NaN value and STD pixels are used.

This bug was reported by Joseph Putko.

This fixes bug #57293.