MakeCatalog: correctly copying unit string by allocation
When the input dataset had a `UNIT' keyword, brightness related columns
like `--brightness' print that unit in their column metadata. Until now,
only a reference to this string was passed to the column, so when the
actual dataset was freed, there would be a double-free crash.

With this commit, when we need such columns, we actually allocate their
necessary space and then actually copy the string's memory, so there will
be no conflict with freeing the input dataset in the end.

This bug was found on a dataset provided by Joseph Putko.

This fixes bug #57300.