MakeCatalog: correctly using average river flux x clump area
Until now, due to some previous mistake in editing MakeCatalog, it was
mistakenly subtracting the total river flux (around each clump) from the
total clump flux when calculating the S/N! But this is wrong!!! It should
be like this: the total clump flux should be divided by the mean river
flux, multiplied by the number of clump pixels used. This resulted in
negative S/N values in some situations.

With this commit, it has been implemented properly.

This bug was found by Joseph Putko.

This fixes bug #57301.