MakeCatalog: column area requested when measuring S/N
Until the previous commit (where the S/N wasn't measured properly!), the
area wasn't used in the calculation of the clump S/N value. But with the
previous commit, the number of pixels in the clump (`CCOL_NUM') is also
used in the clump S/N value.

However, I had forgot to mark this internal column as a dependency of the
S/N calculation. Therefore when the area (or another column that was marked
as needing `CCOL_NUM' like `--mean') wasn't explicitly being requested,
MakeCatalog wouldn't write a value in that column, so it would remain zero,
and thus the clump was effectively being treated as an object!

With this commit, MakeCatalog now knows that the S/N measurement also
depends on the clump area.

This was reported by Joseph Putko.