Fits: new --datasum for print the given HDU's datasum
Until now it was only possible to do this with the `--write=datasum' option
which would actually need write permissions on the input file. But when we
don't have that write permission, it was useless.

With this commit, the Fits program has a new `--datasum' option that will
only need read permission. It will estimate the value and print it on the
command-line (for easy checking in a shell script for example).