FITS Library: using NaN for non-readable float number in ASCII tables
Until now, when an un-readable (by the C Library's `strtod' function)
string was encountered in a floating-point ASCII FITS table's column, the
running program would abort with a crash exactly giving the row and column
number. However, different ASCII FITS table writers, use different
conventions to mark NaNs in ASCII FITS tables and it is hard to include
all.

With this commit, when such strings are encountered, they are read as NaN
values and no warning or error is printed. This is the natural expected
behavior when an un-readable number is present in an ASCII FITS table
column.