NoiseChisel: produces output when there are no detected pixels
Until now, when NoiseChisel didn't detect any pixels, it just printed a
message and wouldn't not make any output file. This was very inconvenient
in general scripts.

With this commit, an output will be created and the binary image will only
have a value of zero. As a result, the Sky and Sky standard deviation
extensions will be measured over all the tiles.