CosmicCalculator: --listlines option to help in remembering line names
Until now, when using CosmicCalculator with spectral lines, it was hard to
remember the exact line name used (you had to go to the proper section of
the manual).

With this commit CosmicCalculator now has the `--listlines' option that
will list all the spectral lines that Gnuastro's spectra library
recognizes. Maybe later as more lines are added, we can specify ranges like
(UV, visible, IR or etc). But for now, there aren't too many.