Book: added explanation for GAL_SPECLINES_INVALID_MAX of last commit
In the previous commit, this new macro was defined in the `speclines.h'
header (with an old name of `GAL_SPECLINES_NUMBER'), but I forgot to put a
description for it in the book. Also, the old name could be misleading, or
confusing (since its not the number of lines, its the number of lines, plus
one).

With this commit, the description is added in the book and also a line is
added to the `NEWS' file, and the name is corrected to this new name which
is more clear. An example usage has also been added in the book.