Book: added nice example of how to limit results of --listlines
While I was playing with the new `--listlines' option of CosmicCalculator,
I noticed a good and useful combination of it with Gnuastro's Table
program, so I added it to the book for users to also use if they need it.

Also, while compiling, I notice the compiler warning that `tdata' in
Table's relatively new `table_selection_equal_or_notequal' function was
extra and never used. It has thus been removed.