Webpage: added section on recent papers that have cited Gnuastro
To demonstration the impact of Gnuastro on the community, 20 of the most
recent papers that cite Gnuastro are now listed on the main English
webpage. I will try to keep it uptodate every once and a while, but there
are links to Google Scholar and ADS for people to get the most recent
results.