Library (reading WCS): warning when WCS values in input are strings
When WCSLIB confronts a string value to a WCS-related keyword, it will put
a zero in its place! But some FITS writers unfortunately place single
quotes around the WCS-related numerical keywords! As a result, WCSLIB isn't
able to read them and will put a zero in their place (causing un-expected
results). Until now, there was no way to check for this.

With this commit, when the primary WCS keywords (`CRVAL*' and `CRPIX*')
have a value of zero, the `gal_wcs_read_fitsptr' function will check to see
if the value indeed starts with a single-quote and if it does, a complete
warning is printed, describing the problem, and informing the user on how
to fix it.