Library (reading WCS): check if lonpole has been inferred properly
When LONPOLE isn't given in the FITS keywords, and `PV%_1' keywords exist
that don't correspond to the points properly, we can end up with absurd
values for `wcs->lonpole' and will given an ambiguous "wcsset ERROR 5:
Invalid parameter value." error.

With this commit, when the value of `wcs->lonpole' isn't reasonable, a
warning will be printed to help the user identify the source of the
problem.

Some minor edits were also made to the warning message of the previous
commit and another arithmetic warning, now properly ends with a new-line.